requirements.txt
streamlit==1.40.2
pandas==2.2.3
numpy==2.2.1
plotly==5.24.1
scikit-learn==1.6.1
streamlit-mic-recorder
SpeechRecognition
twilio
pymongo
torch
torchvision
pillow  # For image handling
requests
python-dotenv
google-generativeai